requirements.txt
sqlalchemy>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0
aiohttp>=3.9.0
httpx>=0.25.0
python-dotenv>=1.0.0
markdown>=3.4.0

